requirements.txt
torch==1.13.0
numpy==1.23.3
tqdm==4.32.2
tensorboard==1.15.0
Pillow==9.2.0
opencv-python